Tests

+++ h2/src/test/org/h2/test/db/TestCases.java
@@ -21,6 +21,7 @@ import java.util.List;
 import java.util.Random;
 
 import org.h2.api.ErrorCode;
+import org.h2.engine.Constants;
 import org.h2.store.fs.FileUtils;
 import org.h2.test.TestBase;
 
@@ -932,20 +933,22 @@ public class TestCases extends TestBase {
         ResultSet rs;
 
         // test the default (SIGNED)
-        stat.execute("create table bin( x binary(1) );");
-        stat.execute("insert into bin(x) values (x'09'),(x'0a'),(x'99'),(x'aa');");
-        rs = stat.executeQuery("select * from bin order by x;");
-        rs.next();
-        assertEquals("99", rs.getString(1));
-        rs.next();
-        assertEquals("aa", rs.getString(1));
-        rs.next();
-        assertEquals("09", rs.getString(1));
-        rs.next();
-        assertEquals("0a", rs.getString(1));
+        if (Constants.VERSION_MINOR < 4) {
+            stat.execute("create table bin( x binary(1) );");
+            stat.execute("insert into bin(x) values (x'09'),(x'0a'),(x'99'),(x'aa');");
+            rs = stat.executeQuery("select * from bin order by x;");
+            rs.next();
+            assertEquals("99", rs.getString(1));
+            rs.next();
+            assertEquals("aa", rs.getString(1));
+            rs.next();
+            assertEquals("09", rs.getString(1));
+            rs.next();
+            assertEquals("0a", rs.getString(1));
+            stat.execute("drop table bin");
+        }
 
         // test UNSIGNED mode
-        stat.execute("drop table bin");
         stat.execute("SET BINARY_COLLATION UNSIGNED");
         stat.execute("create table bin( x binary(1) );");
         stat.execute("insert into bin(x) values (x'09'),(x'0a'),(x'99'),(x'aa');");
